Add support for ascii-armored keyrings
Before this commit, Gradle supported keyrings in binary form (.gpg), which is opaque
and not very handy when performing code reviews. With this commit, we introduce an
option to read a `.keys` file instead, which supports ascii-armored keys.

This means that users can now append keys to the keyring using keys as found in
ascii-armored format, which is for example widely used in the Apache ecosystem.

Such keys can be exported using, for example:

   gpg --armor --export <KEYID>

The file format is lean, meaning that you can have arbitrary text between keys.
This is useful to document the appended keys. For example it is common to
add the uids of the keys in the keyring, for example the output of:

   gpg --list-sigs <KEYID>